[IMP] cetmix_tower_server: Merge cetmix_tower_file into cetmix_tower_server
[REM] cetmix_tower_file
