--- cetmix_tower_file/README.rst
@@ -1,35 +1,0 @@
-**This file is going to be generated by oca-gen-addon-readme.**
-
-*Manual changes will be overwritten.*
-
-Please provide content in the ``readme`` directory:
-
-* **DESCRIPTION.rst** (required)
-* INSTALL.rst (optional)
-* CONFIGURE.rst (optional)
-* **USAGE.rst** (optional, highly recommended)
-* DEVELOP.rst (optional)
-* ROADMAP.rst (optional)
-* HISTORY.rst (optional, recommended)
-* **CONTRIBUTORS.rst** (optional, highly recommended)
-* CREDITS.rst (optional)
-
-Content of this README will also be drawn from the addon manifest,
-from keys such as name, authors, maintainers, development_status,
-and license.
-
-A good, one sentence summary in the manifest is also highly recommended.
-
-
-Automatic changelog generation
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-`HISTORY.rst` can be auto generated using `towncrier <https://pypi.org/project/towncrier>`_.
-
-Just put towncrier compatible changelog fragments into `readme/newsfragments`
-and the changelog file will be automatically generated and updated when a new fragment is added.
-
-Please refer to `towncrier` documentation to know more.
-
-NOTE: the changelog will be automatically generated when using `/ocabot merge $option`.
-If you need to run it manually, refer to `OCA/maintainer-tools README <https://github.com/OCA/maintainer-tools>`_.
