--- cetmix_tower_file/models/__init__.py
@@ -1,5 +1,0 @@
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-
-from . import cx_tower_file
-from . import cx_tower_file_template
-from . import cx_tower_server
