--- cetmix_tower_file/security/ir.model.access.csv
@@ -1,7 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_file_user,File->User,model_cx_tower_file,cetmix_tower_server.group_user,1,0,0,0
-access_file_manager,File->Manager,model_cx_tower_file,cetmix_tower_server.group_manager,1,1,1,0
-access_file_root,File->Root,model_cx_tower_file,cetmix_tower_server.group_root,1,1,1,1
-access_file_template_user,File Template->User,model_cx_tower_file_template,cetmix_tower_server.group_user,1,0,0,0
-access_file_template_manager,File Template->Manager,model_cx_tower_file_template,cetmix_tower_server.group_manager,1,1,1,0
-access_file_template_root,File Template->Root,model_cx_tower_file_template,cetmix_tower_server.group_root,1,1,1,1
