--- cetmix_tower_file/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_cetmix_tower_file
