--- cetmix_tower_file/views/cx_tower_server_view.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-
-    <record id="cx_tower_server_view_form" model="ir.ui.view">
-        <field name="name">cx.tower.server.view.form</field>
-        <field name="model">cx.tower.server</field>
-        <field name="inherit_id" ref="cetmix_tower_server.cx_tower_server_view_form" />
-        <field name="arch" type="xml">
-            <xpath expr="//div[@name='button_box']" position="inside">
-                <button
-                    name="action_open_files"
-                    type="object"
-                    class="oe_stat_button"
-                    icon="fa-pencil-square-o"
-                >
-                    <field name="file_count" widget="statinfo" string="Files" />
-                </button>
-            </xpath>
-        </field>
-    </record>
-
-</odoo>
