--- cetmix_tower_file/views/menuitems.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-
-    <menuitem
-        id="menu_cx_tower_file_root"
-        name="Files"
-        parent="cetmix_tower_server.menu_root"
-        groups="cetmix_tower_server.group_user"
-        sequence="4"
-    />
-    <menuitem
-        id="menu_cx_tower_file"
-        name="Files"
-        action="cx_tower_file_action"
-        parent="menu_cx_tower_file_root"
-        sequence="1"
-    />
-    <menuitem
-        id="menu_cx_tower_file_template"
-        name="Templates"
-        action="cx_tower_file_template_action"
-        parent="menu_cx_tower_file_root"
-        sequence="2"
-    />
-
-</odoo>
