+++ cetmix_tower_server/models/cx_tower_server.py
@@ -1,5 +1,6 @@
 # Copyright (C) 2022 Cetmix OÜ
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+import ast
 import io
 import logging
 
@@ -269,6 +270,20 @@ class CxTowerServer(models.Model):
     plan_log_ids = fields.One2many(
         comodel_name="cx.tower.plan.log", inverse_name="server_id"
     )
+    file_ids = fields.One2many(
+        "cx.tower.file",
+        "server_id",
+        string="Files",
+    )
+    file_count = fields.Integer(
+        "Total Files",
+        compute="_compute_file_count",
+    )
+
+    def _compute_file_count(self):
+        """Compute total server files"""
+        for server in self:
+            server.file_count = len(server.file_ids)
 
     @api.constrains("ip_v4_address", "ip_v6_address", "ssh_auth_mode")
     def _constraint_ssh_settings(self):
@@ -746,3 +761,26 @@ class CxTowerServer(models.Model):
                 _('The file "{}" not found.'.format(remote_path))
             ) from fe
         return result
+
+    def action_open_files(self):
+        """
+        Open current server files
+        """
+        action = self.env["ir.actions.actions"]._for_xml_id(
+            "cetmix_tower_server.cx_tower_file_action"
+        )
+        action["domain"] = [("server_id", "=", self.id)]
+
+        context = self._context.copy()
+        if "context" in action and isinstance((action["context"]), str):
+            context.update(ast.literal_eval(action["context"]))
+        else:
+            context.update(action.get("context", {}))
+
+        context.update(
+            {
+                "default_server_id": self.id,
+            }
+        )
+        action["context"] = context
+        return action
