+++ cetmix_tower_server/tests/__init__.py
@@ -3,3 +3,4 @@ from . import test_variable
 from . import test_command
 from . import test_key
 from . import test_plan
+from . import test_file
