+++ cetmix_tower_server/views/cx_tower_server_view.xml
@@ -53,7 +53,14 @@
                             class="oe_stat_button"
                             icon="fa-bars"
                         />
-
+                        <button
+                            name="action_open_files"
+                            type="object"
+                            class="oe_stat_button"
+                            icon="fa-pencil-square-o"
+                        >
+                            <field name="file_count" widget="statinfo" string="Files" />
+                        </button>
                     </div>
                     <div class="oe_title">
                         <h1>
