+++ setup/_metapackage/setup.py
@@ -8,7 +8,7 @@ setuptools.setup(
     description="Meta package for cetmix-cetmix-tower Odoo addons",
     version=version,
     install_requires=[
-        'odoo14-addon-cetmix_tower_file',
+        # 'odoo14-addon-cetmix_tower_file',
         'odoo14-addon-cetmix_tower_server',
         'odoo14-addon-cetmix_tower_server_queue',
     ],
