--- setup/cetmix_tower_file/odoo/addons/cetmix_tower_file
@@ -1,1 +1,0 @@
-../../../../cetmix_tower_file
